Moved DI logic in dedicated folder
